+++ frontend/lib/components/navigation_ui.dart
@@ -1,0 +1,30 @@
+import "package:flutter/material.dart";
+import "package:google_maps_flutter/google_maps_flutter.dart";
+
+class NavigationUI extends StatefulWidget {
+  const NavigationUI({super.key});
+
+  @override
+  State<NavigationUI> createState() => _NavigationUIState();
+}
+
+class _NavigationUIState extends State<NavigationUI> {
+  late GoogleMapController mapController;
+
+  final LatLng _center = const LatLng(45.521563, -122.677433);
+
+  void _onMapCreated(GoogleMapController controller) {
+    mapController = controller;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GoogleMap(
+      onMapCreated: _onMapCreated,
+      initialCameraPosition: CameraPosition(
+        target: _center,
+        zoom: 11.0,
+      ),
+    );
+  }
+}
